swupdate: Preserve files when writing new image
Process a list of files/directories from /etc/swu-transfer-list and copy
them to the newly installed image from the running image. By default
preserve the hostname and dropbear's host key. The list can be edited
by a user as a config file. It's also quite possible to add the file
to itself so that the list is kept too.

Only the SD/eMMC modes currently call the script.