Use kexec kernel to upgrade running rootfs in QSPI flash
When there's only one partition, it cannot be upgraded while running.
Solve this by starting a new image from RAM, using kexec, and use the
image in RAM to flash the QSPI contents.
When not running from the same partition, the kexec step is skipped.
This allows a quicker update when booted from SD.

This also demonstrates how one could send a test image to the board
to boot in RAM directly.