xdpzu7: Fix WiFi not working and handle SD card boot
The WiFi and SD card on the xdpzu7 share the same controller so only
one can really be active. Supply separate devicetrees so that the
WiFi is kept in reset when the SD card is active. Active the SD
devicetree on the SD card using the ROOTFS_FIXUP_COMMANDS hook.

This fixes the WiFi not working, since the WiFi is "behind" the FPGA
so it has to wait for the FPGA to be programmed before starting up.