swupdate: cleanup & make it possible to include swu images in other images
It was not possible to include SWU images in another rootfs. By solving
the manifest errors and dependencies on the rootfs image it becomes
possible to include the image in the factory programming images.

Also remove a sync in partition_sd_card which could trigger an non-zero
exit code because the kernel is rereading the partition table in the
background.

[ML] Rebased on master, removed extra "." in README.md