ifplugd-auto-net: Add support for WEXT based wifi modules
Some older modules still use the WEXT (wifi extensions) interface. There is support for
this in wpa_supplicant, but it has to be enabled by passing the interfaces on the command
line. Add "wext" to the driver list, this makes wpa_supplicant attempt "nl80211" first, and
if that fails it tries "wext". Thus both interfaces work after this change.