busybox: Support hotplug for network devices
When a network device is attached (e.g. wifi dongle or usb network adapter) while the system is
running, the "ifplugd" deamon was not started for it automatically.

Modify mdev.conf to listen to network devices (ethX, wlanX, usbX) and call the mdev-network.sh
script when a device has been added or removed. The mdev-network.sh script then starts and
stops the ifplugd deamon if that has heen configured on the system. Add an "flock" mutex to
the scripts so that only one instance per interface is ever started.

Thus, inserting a WiFi or ethernet dongle now makes it start up and initialize fully.

This also fixes that the usb gadget ethernet would stop working after the OTG port has been
used in host mode. Now it will re-initialize when the USB OTG switches back to peripheral mode.