swu-image: Run "git rev-parse" commands outside fakeroot
git recently fixed a security issue:
  https://github.blog/2022-04-12-git-security-vulnerability-announced/

This results in builds failing with an error like:

  Exception: bb.process.ExecutionError: Execution of '.../temp/run.rootfs_add_swu_metadata.7743' failed with exit code 128:
  fatal: unsafe repository ('.../...-platform' is owned by someone else)
  To add an exception for this directory, call:
	git config --global --add safe.directory .../...-platform
  WARNING: .../temp/run.rootfs_add_swu_metadata.7743:1 exit 128 from 'v=`git rev-parse --verify HEAD`'

The problem is that do_rootfs runs under fakeroot, and git thus thinks we're
actually 'root' while executing that rev-parse command. To solve the issue,
insert a task that runs outside fakeroot and creates the files.

Also move the IMAGEVERSION variable outside the "swu.inc" as only "swu-image.inc"
requires this variable, and it triggers the same issue otherwise.