install_to_sd: Support writing raw partitions
Support writing a raw ext4 (or other) filesystem to the SD card.

Remove support for obsolete options like FPGA_BOOT_IMAGE and
COPY_LOADABLE_MODULES.

Expects a ROOTFSTYPE variable to specify the filesystem type as also
used in OE, most commonly used would be "ext4.gz". It defaults to tar.gz
which mimics the old behavior. Other filesystems are copied directly
to the active partition using "dd". The script attempts to resize and
rename the partition afterwards.

Since the partition name may not be "sd-rootfs-X", in particular when
the filesystem is squashfs, allow specifying the device name as a
commandline parameter.