poweroff-key: Program to shut down system on KEY_POWER
When running systemd, the KEY_POWER button is handled internally and the
system gracefully shuts down when a user presses that key.
This program provides that functionality when running initscripts for
example. It looks for a keyboard that has a KEY_POWER key, and if found,
waits for that key to be pressed and shuts down the system. If there's
no such key, the program will simply exit.