expand-wic-partition: Repair and limit rootfs sizes
Newer parted versions have an obscure bug that it doesn't process the
remainder of the commandline when it wanted to ask an "are you sure"
question. Work around this by calling parted multiple times.

Using 40% as rootfs size made sense when SD cards were 2GB to 8GB in size,
but now that they are typically 16GB or more, this results in creating
huge partitions for the rootfs copies. Check the device's size, and when
it's larger than 5GB, just create root filesystem partitions of just under
2GB each, such that the "data" partition aligns on a GiB boundary.

Apply similar changes to partition_sd_card.sh, which is also being used
to initialize on-board eMMC devices.