images: Remove board filter on init images and make QSPI init work again
The init images are often applied to many boards as they only really affect
the SOM. So remove the "board" filter and allow this to be installed on
any board.

The QSPI "spl" mode did not work, apparently swupdate fixed the bug that
it erased the whole device, so now we should provide type="flash" for all
binaries that are to be written into the boot partition of the MTD device.