images: Support a read-only squashfs-lz4 image
Add "my-image-ro" that builds a read-only rootfs using squashfs-lz4. This
image can be updated as usual through SWU, or just written to SD card.
If the board supports it (all topic boards do), builds a WIC image too.

This offers a fast and small image suitable for production environments.
The my-image-ro.bb provides a template for other projects.

It should be possible to write the squashfs image to an UBI volume and
boot this directly from within u-boot. This is still under investigation.
Since my-image-ro is below 13MB for the tdkz30 target, it would allow an
A/B update system in a 32MB QSPI flash configuration.