images: Don't let avahi-deamon depend on package-management feature
Having avahi is also useful for SWUpdate and other network applications.
It doesn't make sense to only include avahi-deamon into an image when it
has the "package-management" feature enabled. Instead, create an
MDNS_PROVIDER variable and always include that in the image. This allows
projects to remove it by setting MDNS_PROVIDER="".